add tests for importing prototyped functions
Unfortunately this butts up against a limitation of Perl, which is that
functions with BLOCK prototype need to be seen by the parser during compile
time. If you try to dynamically load a function with a BLOCK prototype at
runtime, the parse doesn't understand how to treat the BLOCK argument.

Specifically, this fails:

    #!/usr/bin/env perl

    use strict;
    use warnings;

    sub make_function {
      eval 'sub with_proto (&@) { my $c = shift; return map { $c->() } @_ }';
      die $@ if $@;
      print "$_\n" foreach with_proto { uc($_) } qw(a e i o u);
    }

    make_function();
    __END__
    Can't call method "with_proto" without a package or object reference at - line 9.

Whereas applying this change:

    @@ -4,8 +4,10 @@
     use warnings;

     sub make_function {
    -  eval 'sub with_proto (&@) { my $c = shift; return map { $c->() } @_ }';
    -  die $@ if $@;
    +  BEGIN {
    +    eval 'sub with_proto (&@) { my $c = shift; return map { $c->() } @_ }';
    +    die $@ if $@;
    +  }
       print "$_\n" foreach with_proto { uc($_) } qw(a e i o u);
     }

causes the correct output: A E I O U

Other prototypes seem not to have this problem; presumably because such function calls are
less ambiguous to the parser than `sub func BLOCK` is.

Since the entire point of this module is to defer loading other code to
runtime, changing this code:

  sub do_something {
    my ($class, @args) = @_;

    $class->dynamic_use('func_with_block_prototype')->class_that_provides_that_function;
    func_with_block_prototype { ... } @args;
  }

to load 'func_with_prototype' at BEGIN time:

  sub do_something {
    my ($class, @args) = @_;

    BEGIN {
      $class->dynamic_use('func_with_block_prototype')->class_that_provides_that_function;
    }

    func_with_block_prototype { ... } @args;
  }

is probably not desirable. This commit illustrates that bypassing the
prototype allows the entire process to happen at runtime:

  sub do_something {
    my ($class, @args) = @_;

    $class->dynamic_use('func_with_block_prototype')->class_that_provides_that_function;
    func_with_block_prototype(sub { ... }, @args);
  }

Unfortunately, the '{;' method of disambiguating HASHREF-vs-BLOCK mentioned
in `perldoc perlref` does not work for getting runtime-loaded functions
with BLOCK prototypes to compile correctly.

So:
  - if you're importing stuff at runtime, it mostly works
  - unless the thing you're importing has a BLOCK prototype
  - and you're trying to call that prototyped function with a BLOCK
    rather than a CODEREF as its argument
  - in which case you need to provide a CODEREF, not a BLOCK