`use universal::dynamic_use qw(Foo::prototyped)` does something
Perl doesn't allow you to create BLOCK-prototyped functions at runtime.
However, if we give a hint to this module, and assure it that we really
do think we know what we're doing, then `universal::dynamic_use` will
try *really hard* at compile-time to track down the prototype for a
given function; if a prototype is found, then `universal::dynamic_use`
will forward-declare a subroutine in the calling package with the
correct prototype.

This forward-declaration allows the Perl parser to set itself up correctly
to understand that a call such as this:

    my @uc = apply_block { uc($_) } ('a'..'f');

is to be understood as

    my @uc = apply_block(sub { uc($_) }, ('a'..'f'));

rather than the invalid:

    my @uc = +{ uc($_) }->apply_block('a'..'f');

which seems to be what's going on when we try declaring BLOCK-prototyped
functions at runtime.

This code is sufficiently crazy that I put it behind a double firewall of
warning messages.