non-overridable CORE functions now trigger compile errors
Any Perl function that has CORE::prototype of undef is a non-overridable
function. Well, actually, that's not quite true; that's just what
'perldoc -f prototype' woudld have you believe.

A full list of perl's built-in functions that have no prototype can be
generated from this command:

  $ perldoc -m perlfunc | perl -lne '
      next unless /Perl Functions by Category/ .. /Portability/;
      if (/[IC]</) {
        push @f, map {
          s/\s+//;
          s/[IC]<//g;
          s/>//g;
          s{[^a-zA-Z0-9_\-/]}{}g;
          $_
        }
        grep { /[IC]</ }
        map { split / /, $_ }
        split /,/, $_;
      }

      END {
        %s;
        print foreach
          grep { ! $s{$_}++ && ! defined eval { CORE::prototype("CORE::$_") } }
          map { s,/,,g; $_ }
          sort {
            ($b =~ m{/}) <=> ($a =~ m{/}) ||
            $a cmp $b
          } @f;
      }
  '

For perl v5.10.1, this list is:

    m//, q//, qq//, qr//, qw//, qx//, s///, tr///, y///, -X,
    chomp, chop, default, defined, delete, do, eval, exec,
    exists, format, given, glob, goto, grep, import, last,
    local, map, my, next, no, our, package, pos, print,
    printf, prototype, redo, require, return, say, scalar,
    sort, split, state, study, sub, system, tie, tied, undef,
    untie, use, when

However, this list is incomplete ('if' doesn't appear there, because it
is not a core function; it is a keyword) and it is inaccurate ('glob'
is well-known to be overridable despite having no prototype).

Maybe in another commit I'll extract all the toke.c:Perl_keywords() from
each version of Perl in order to allow this module to truly do what it's
intending to do by checking the definedness of the prototype.